requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# Streamlit Frontend
streamlit==1.29.0

# PDF/DOCX Processing
PyMuPDF==1.23.8
python-docx==1.1.0

# NLP and ML
spacy==3.7.2
sentence-transformers==2.7.0
transformers==4.39.3
huggingface-hub==0.23.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# LLM Integration
google-generativeai==0.6.0

# Database
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
requests==2.31.0
